@@ -20,57 +20,56 @@
     "prepare": "npm run build && ionic cordova prepare"
   },
   "dependencies": {
-    "@angular/animations": "^7.2.12",
-    "@angular/cdk": "^7.3.6",
-    "@angular/common": "^7.2.12",
-    "@angular/core": "^7.2.12",
-    "@angular/forms": "^7.2.12",
-    "@angular/http": "^7.2.12",
-    "@angular/material": "^7.3.6",
-    "@angular/platform-browser": "^7.2.12",
-    "@angular/platform-browser-dynamic": "^7.2.12",
-    "@angular/router": "^7.2.12",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
     "@angular/tsc-wrapped": "^4.4.6",
-    "@ionic-native/android-permissions": "^5.3.0",
-    "@ionic-native/core": "^5.3.0",
-    "@ionic-native/http": "^5.3.0",
-    "@ionic-native/splash-screen": "^5.3.0",
-    "@ionic-native/status-bar": "^5.3.0",
-    "@ionic/angular": "^4.2.0",
-    "@ionic/angular-toolkit": "^1.5.0",
-    "cordova-android": "^8.0.0",
-    "cordova-plugin-advanced-http": "^2.0.7",
-    "cordova-plugin-android-permissions": "^1.0.0",
-    "cordova-plugin-device": "^2.0.2",
-    "cordova-plugin-file": "^6.0.1",
-    "cordova-plugin-ionic-keyboard": "^2.1.3",
-    "cordova-plugin-ionic-webview": "^4.0.1",
-    "cordova-plugin-splashscreen": "^5.0.2",
-    "cordova-plugin-statusbar": "2.4.2",
-    "cordova-plugin-whitelist": "^1.3.3",
-    "core-js": "^3.0.0",
-    "rxjs": "^6.4.0",
-    "rxjs-compat": "^6.4.0",
-    "zone.js": "^0.9.0"
+    "@ionic-native/android-permissions": "^5.19.1",
+    "@ionic-native/core": "^5.19.1",
+    "@ionic-native/http": "^5.19.1",
+    "@ionic-native/splash-screen": "^5.19.1",
+    "@ionic-native/status-bar": "^5.19.1",
+    "@ionic/angular": "^4.11.8",
+    "@ionic/angular-toolkit": "^2.1.2",
+    "cordova-android": "^8.1.0",
+    "cordova-plugin-advanced-http": "^2.3.1",
+    "cordova-plugin-android-permissions": "^1.0.2",
+    "cordova-plugin-device": "^2.0.3",
+    "cordova-plugin-file": "^6.0.2",
+    "cordova-plugin-ionic-keyboard": "^2.2.0",
+    "cordova-plugin-ionic-webview": "^4.1.3",
+    "cordova-plugin-splashscreen": "^5.0.3",
+    "cordova-plugin-statusbar": "2.4.3",
+    "cordova-plugin-whitelist": "^1.3.4",
+    "core-js": "^3.6.4",
+    "rxjs": "^6.5.4",
+    "rxjs-compat": "^6.5.4",
+    "zone.js": "^0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular-devkit/core": "^7.3.8",
-    "@angular/cli": "^7.3.8",
-    "@angular/compiler": "^7.2.12",
-    "@angular/compiler-cli": "^7.2.12",
-    "@ionic/lab": "^1.0.24",
-    "@types/jasmine": "^3.3.12",
-    "@types/node": "^11.13.0",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "^3.4.0",
-    "karma": "^4.0.1",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
-    "karma-jasmine": "^2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
+    "@angular-devkit/build-angular": "^0.803.23",
+    "@angular-devkit/core": "^8.3.23",
+    "@angular/cli": "^8.3.23",
+    "@angular/compiler": "^8.2.14",
+    "@angular/compiler-cli": "^8.2.14",
+    "@ionic/lab": "^2.0.19",
+    "@types/jasmine": "^3.5.1",
+    "@types/node": "^13.1.8",
+    "codelyzer": "^5.2.1",
+    "jasmine-core": "^3.5.0",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
+    "karma-jasmine": "^3.1.0",
+    "karma-jasmine-html-reporter": "^1.5.1",
     "tslint": "^5.15.0",
-    "typescript": "^3.2.4"
+    "typescript": "^3.5.3"
   },
   "cordova": {
     "plugins": {
